Reorganize dtype and common code for easier maintainability
