+++ gs/numpy/random.py
@@ -1,12 +1,12 @@
 """Numpy based random backend."""
 
 from ._dispatch import numpy as _np
-from ._dispatch._common import (
+from ._dispatch.numpy.random import default_rng as _default_rng  # noqa: F401
+from ._dispatch.numpy.random import randint, seed  # noqa: F401
+from ._dtype import (
     _allow_complex_dtype,
     _modify_func_default_dtype,
 )
-from ._dispatch.numpy.random import default_rng as _default_rng
-from ._dispatch.numpy.random import randint, seed
 
 rand = _modify_func_default_dtype(
     copy=False, kw_only=True, target=_allow_complex_dtype(target=_np.random.rand)
