+++ gs/pytorch/_common.py
@@ -1,4 +1,3 @@
-import numpy as _np
 import torch as _torch
 
 
@@ -10,9 +9,3 @@ def from_numpy(x):
 
 def array(val, dtype=None):
     return _torch.asarray(val, dtype=dtype)
-
-
-def cast(x, dtype):
-    if _torch.is_tensor(x):
-        return x.to(dtype=dtype)
-    return array(x, dtype=dtype)
