+++ tests/conftest.py
@@ -1,4 +1,4 @@
-from geomstats.test.conf import (
+from geomstats.test.conf import (  # noqa: F401
     pytest_collection_modifyitems,
     pytest_make_parametrize_id,
 )
